Organize Kira integration, make sliders update the EQ effect
