+++ crates/phonon-kira/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 phonon = { path = "../phonon" }
 kira = "0.8"
+ringbuf = "0.3"
 
 [dev-dependencies]
 eframe = "0.27"
