+++ crates/phonon-kira/examples/eq_effect.rs
@@ -1,19 +1,18 @@
-use eframe::egui::Context;
-use eframe::{egui, Frame};
+use eframe::egui;
 use kira::manager::backend::cpal::CpalBackend;
 use kira::manager::{AudioManager, AudioManagerSettings};
 use kira::sound::static_sound::{StaticSoundData, StaticSoundSettings};
 use kira::track::TrackBuilder;
-use phonon_kira::EqEffectBuilder;
+use phonon_kira::eq_effect::builder::EqEffectBuilder;
 
 fn main() {
-    let eq_gains: [f32; 3] = [1.0, 0.5, 1.0];
+    let mut eq_gains: [f32; 3] = [1.0, 1.0, 1.0];
 
     let mut manager = AudioManager::<CpalBackend>::new(AudioManagerSettings::default()).unwrap();
 
-    let track = manager
-        .add_sub_track(TrackBuilder::new().with_effect(EqEffectBuilder { eq_gains }))
-        .unwrap();
+    let mut track_builder = TrackBuilder::new();
+    let mut effect_handle = track_builder.add_effect(EqEffectBuilder { eq_gains });
+    let track = manager.add_sub_track(track_builder).unwrap();
 
     let sound_data = StaticSoundData::from_file(
         "data/audio/pink_noise.ogg",
@@ -23,36 +22,18 @@ fn main() {
 
     manager.play(sound_data).unwrap();
 
-    eframe::run_native(
-        "EQ Effect Example (Kira)",
+    eframe::run_simple_native(
+        "EQ Effect (Kira)",
         eframe::NativeOptions::default(),
-        Box::new(|cc| Box::<App>::default()),
+        move |ctx, _frame| {
+            egui::CentralPanel::default().show(ctx, |ui| {
+                ui.add(egui::Slider::new(&mut eq_gains[0], 0.0..=1.0).text("Gain Low"));
+                ui.add(egui::Slider::new(&mut eq_gains[1], 0.0..=1.0).text("Gain Mid"));
+                ui.add(egui::Slider::new(&mut eq_gains[2], 0.0..=1.0).text("Gain High"));
+            });
+
+            effect_handle.set_gains(eq_gains).unwrap();
+        },
     )
-    .unwrap();
-}
-
-struct App {
-    eq_gain0: f32,
-    eq_gain1: f32,
-    eq_gain2: f32,
-}
-
-impl Default for App {
-    fn default() -> Self {
-        Self {
-            eq_gain0: 1.0,
-            eq_gain1: 1.0,
-            eq_gain2: 1.0,
-        }
-    }
-}
-
-impl eframe::App for App {
-    fn update(&mut self, ctx: &Context, frame: &mut Frame) {
-        egui::CentralPanel::default().show(ctx, |ui| {
-            ui.add(egui::Slider::new(&mut self.eq_gain0, 0.0..=1.0).text("Gain Low"));
-            ui.add(egui::Slider::new(&mut self.eq_gain1, 0.0..=1.0).text("Gain Mid"));
-            ui.add(egui::Slider::new(&mut self.eq_gain2, 0.0..=1.0).text("Gain High"));
-        });
-    }
+    .unwrap()
 }
