+++ crates/phonon-kira/src/eq_effect/mod.rs
@@ -1,0 +1,7 @@
+pub mod builder;
+pub(crate) mod effect;
+pub mod handle;
+
+pub(crate) enum Command {
+    SetGains([f32; 3]),
+}
